requirements.txt
streamlit
pandas
mysql-connector-python
altair
openpyxl
